backend/src/user_create/requirements.txt
psycopg2-binary==2.9.1
redis==4.0.3